backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
firebase-admin==6.2.0
transformers==4.35.2
torch==2.7.0
pillow==10.1.0
scikit-learn==1.3.2
pandas==2.1.2
httpx==0.25.0
pytest==7.4.3
pytest-asyncio==0.21.1
aiohttp==3.9.0
